backend/auth-service/requirements.txt
# File: backend/auth-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
PyJWT==2.8.1
pytest==7.4.4
pytest-asyncio==0.23.2
httpx==0.25.2
pytest-cov==4.1.0
python-json-logger==2.0.7
black==23.12.0
isort==5.13.2
flake8==6.1.0
mypy==1.8.0
python-dotenv==1.0.0
email-validator==2.1.0
